+++ src/Entity/Client.php
@@ -8,15 +8,27 @@ class Client extends Entity {
 	 * @param  string  $id
 	 * @param  string  $secret
 	 * @param  string  $name
+	 * @param  bool  $trusted
 	 * @param  string  $redirectUri
 	 * @return void
 	 */
-	public function __construct($id, $secret, $name, $redirectUri = null)
+	public function __construct($id, $secret, $name, $trusted, $redirectUri = null)
 	{
 		$this->id = $id;
 		$this->secret = $secret;
 		$this->name = $name;
+		$this->trusted = $trusted;
 		$this->redirectUri = $redirectUri;
+	}
+
+	/**
+	 * Check if a client is trusted.
+	 * 
+	 * @return bool
+	 */
+	public function isTrusted()
+	{
+		return $this->trusted == true;
 	}
 
 }