+++ tests/GrantPasswordTest.php
@@ -90,7 +90,7 @@ class GrantPasswordTest extends PHPUnit_Framework_TestCase {
 		});
 
 		$storage->shouldReceive('get')->with('client')->andReturn(m::mock([
-			'get' => new ClientEntity('test', 'test', 'test')
+			'get' => new ClientEntity('test', 'test', 'test', false)
 		]));
 
 		$storage->shouldReceive('get')->with('token')->andReturn(m::mock([